spotify_auth.py
from spotipy import Spotify
from spotipy.oauth2 import SpotifyOAuth
from dotenv import load_dotenv
import os

load_dotenv()

sp = Spotify(auth_manager=SpotifyOAuth(
    client_id=os.getenv("SPOTIPY_CLIENT_ID"),
    client_secret=os.getenv("SPOTIPY_CLIENT_SECRET"),
    redirect_uri="https://example.com/callback",
    scope="user-modify-playback-state user-read-playback-state user-read-currently-playing",
    open_browser=True
))

print("Connected!")


